+++ src/Config/Option/QrCode/DefaultSizeConfigOption.php
@@ -5,13 +5,11 @@ declare(strict_types=1);
 namespace Shlinkio\Shlink\Installer\Config\Option\QrCode;
 
 use Shlinkio\Shlink\Installer\Config\Option\BaseConfigOption;
-use Shlinkio\Shlink\Installer\Config\Util\ConfigOptionsValidatorsTrait;
+use Shlinkio\Shlink\Installer\Config\Util\ConfigOptionsValidator;
 use Symfony\Component\Console\Style\StyleInterface;
 
 class DefaultSizeConfigOption extends BaseConfigOption
 {
-    use ConfigOptionsValidatorsTrait;
-
     private const MIN_SIZE = 50;
     private const MAX_SIZE = 1000;
 
@@ -25,7 +23,7 @@ class DefaultSizeConfigOption extends BaseConfigOption
         return $io->ask(
             'What\'s the default size, in pixels, you want generated QR codes to have (50 to 1000)',
             '300',
-            fn (mixed $value) => $this->validateNumberBetween($value, self::MIN_SIZE, self::MAX_SIZE),
+            fn (mixed $value) => ConfigOptionsValidator::validateNumberBetween($value, self::MIN_SIZE, self::MAX_SIZE),
         );
     }
 }