+++ src/Config/Option/RabbitMq/RabbitMqPortConfigOption.php
@@ -4,13 +4,11 @@ declare(strict_types=1);
 
 namespace Shlinkio\Shlink\Installer\Config\Option\RabbitMq;
 
-use Shlinkio\Shlink\Installer\Config\Util\ConfigOptionsValidatorsTrait;
+use Shlinkio\Shlink\Installer\Config\Util\ConfigOptionsValidator;
 use Symfony\Component\Console\Style\StyleInterface;
 
 class RabbitMqPortConfigOption extends AbstractRabbitMqEnabledConfigOption
 {
-    use ConfigOptionsValidatorsTrait;
-
     public function getEnvVar(): string
     {
         return 'RABBITMQ_PORT';
@@ -22,7 +20,7 @@ class RabbitMqPortConfigOption extends AbstractRabbitMqEnabledConfigOption
         return (int) $io->ask(
             'RabbitMQ port',
             $useSsl ? '5671' : '5672',
-            fn (mixed $value) => $this->validateNumberBetween($value, 1, 65535),
+            fn (mixed $value) => ConfigOptionsValidator::validateNumberBetween($value, 1, 65535),
         );
     }