+++ src/Config/Option/Visit/VisitsThresholdConfigOption.php
@@ -5,13 +5,11 @@ declare(strict_types=1);
 namespace Shlinkio\Shlink\Installer\Config\Option\Visit;
 
 use Shlinkio\Shlink\Installer\Config\Option\BaseConfigOption;
-use Shlinkio\Shlink\Installer\Config\Util\ConfigOptionsValidatorsTrait;
+use Shlinkio\Shlink\Installer\Config\Util\ConfigOptionsValidator;
 use Symfony\Component\Console\Style\StyleInterface;
 
 class VisitsThresholdConfigOption extends BaseConfigOption
 {
-    use ConfigOptionsValidatorsTrait;
-
     public function getEnvVar(): string
     {
         return 'DELETE_SHORT_URL_THRESHOLD';
@@ -23,7 +21,7 @@ class VisitsThresholdConfigOption extends BaseConfigOption
             'What is the amount of visits from which the system will not allow short URLs to be deleted? Leave empty '
             . 'to always allow deleting short URLs, no matter what',
             null,
-            [$this, 'validateOptionalPositiveNumber'],
+            ConfigOptionsValidator::validateOptionalPositiveNumber(...),
         );
 
         return $result === null ? null : (int) $result;